relocate gather channel func
